WIP: compose: Add `experimental/sysusers` option
First, we add support for a new `experimental:` key.  Then
there's a new `sysusers` key underneath that.

When enabled, we drop all of the other previous passwd handling.
In practice the only one that was used was having static files.
That is a pain to maintain.

However, we need to statically assign non-zero uid/gid for any
files that come from a base ostree commit.  Anything else
would mean the user/groups could be unpredictably assigned in different
rpm-ostree runs.

This code now checks for an errors out on that.

In order to convert *fully* to sysusers, we install an interceptor for
`useradd/groupadd` that talk back via a pipe to the compose process.
These invocations then get translated to drop into a new`
`sysusers.d/rpmostree-auto.conf` file.

This way we don't need to require that every RPM have ported to
sysusers.d.

At the end, we drop everything in `/etc/passwd` and `/etc/group`
except for the `root:` entries, relying on `systemd-sysusers` to
readd everything at boot time.

Closes: https://github.com/projectatomic/rpm-ostree/issues/49