daemon: Add new ModifyYumRepo DBus API
This allows clients such as gnome-software to enable and disable
yum repositories.

The API is generic, but for now we only allow changing the 'enabled'
key. If needed, it's easy to allow changing other settings in the
future. See the discussion in the PR for the reasoning.

Closes: #1771

Closes: #1780
Approved by: cgwalters