compose: Delete non-unified-core code
We still accept the `--unified-core` option, it just does
nothing.

We're talking about making 2019.2 the last release to support
RHEL7.  That makes it also a good cutover point to just drop
the non-unified-core codepath.

Since coreos-assembler defaults to it, we know it works.
I also tested Silverblue with it.

Alternative to https://github.com/projectatomic/rpm-ostree/pull/1739