composetree: allow url in manifest package list
This allows URLs to be used in the manifest file when doing a
treecompose only in unified-core mode. This is a convenient way of
adding one off packages for testing purposes but should not be the
preferred method of composing a tree.

The import local rpm functions were also moved to rpmostree-importer
to facilitate usage.