WIP: Add /etc/rpm-ostree/roothooks.d
This is intended to support kernel module systems like
[atomic-wireguard](https://github.com/jdoss/atomic-wireguard).
See the Fedora CoreOS tracker issue:
https://github.com/coreos/fedora-coreos-tracker/issues/249

With a "roothook", one can perform arbitrary modifications to the *new* root
filesystem; if a hook exits with an error, that also stops the upgrade.

Specifically with this, atomic-wireguard could *block* an upgrade if
the new kernel isn't compatible with a module.