Remove `make vmcheck` wrapper
Using `make` as an entrypoint works best when the extended target
needs to interact somehow with building the software.
We need `make check` to be a target becuase it actually does
build code.

But for `vmcheck` that's not true.  And in particular using
`make` targets is awkward when one wants
to configure/shortcut what the target is doing, as we have with
the environment variables `SKIP_VMOVERLAY` etc.

I think it's just cleaner and more obvious to have a script to
execute - and this script can have command line arguments.

All this does is remove the `make vmcheck` wrapper, but I
want to rewrite the entrypoint script to have e.g.
`./tests/vmcheck -O` where `-O` means "skip overlay".