Add a CHANGELOG.md
Different reasons for this, this website elaborates a bit on them:

https://keepachangelog.com/

But there's more to it from my POV. This will make releases easier, and
I think one step closer towards automating it without losing the "human
touch" of release notes.

Now, we could make this part of commit messages and e.g. extract them.
This is the approach taken by
[conventional-changelog](https://github.com/conventional-changelog/standard-version).

Though I'm not entirely convinced with that approach. For one, commit
messages are for a different audience. Second, changelog entries are
meant to be much less detailed than commits and may be represented by
multiple commits/PRs.

Anyway, if we do this, I think we should start with the easy step first
of doing essentially what we already do during releases at PR time.
Then we can look at streamlining it if needed.